backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.4.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic-core==2.14.1
python-multipart==0.0.6
httpx==0.25.2
